Changed GUI to display that connection has been made "Still has bugs"
This commit prints out that a connection has been made if the playerID
is not 1. There are some issues. One is that I do not know if the
connection is actually being made over the server or if a value is just
being assigned. Also the first player to connect to the server has a
value of 0 so it is not tripping and saying that the connection has been
made. This needs to be fixed in the next commit